Rename codec to stream
