+++ subprojects/lucene/src/test/scala/at/bioinform/lucene/DnaTokenizerTest.scala
@@ -1,4 +1,4 @@
-package at.bioinform.codec
+package at.bioinform.stream
 
 import java.io.StringReader
