+++ subprojects/stream/src/main/scala/at/bioinform/stream/fasta/FastaFlow.scala
@@ -1,4 +1,4 @@
-package at.bioinform.codec.fasta
+package at.bioinform.stream.fasta
 
 import java.net.URI
 import java.nio.file.{Path, Paths}
