+++ subprojects/stream/src/main/scala/at/bioinform/stream/fasta/FastaHeader.scala
@@ -1,3 +1,3 @@
-package at.bioinform.codec.fasta
+package at.bioinform.stream.fasta
 
 case class FastaHeader(id: String, description: Option[String])
