+++ subprojects/stream/src/main/scala/at/bioinform/stream/lucene/LuceneSink.scala
@@ -1,8 +1,8 @@
-package at.bioinform.codec.lucene
+package at.bioinform.stream.lucene
 
 import akka.stream.stage.{GraphStageLogic, GraphStageWithMaterializedValue, InHandler}
 import akka.stream.{Attributes, Inlet, SinkShape}
-import at.bioinform.codec.fasta.{FastaEntry, Splitter}
+import at.bioinform.stream.fasta.{FastaEntry, Splitter}
 import at.bioinform.lucene.Util
 import org.apache.lucene.document.Document
 import org.apache.lucene.index.{IndexWriter, IndexWriterConfig}
@@ -19,7 +19,7 @@ import scala.concurrent.{Future, Promise}
  * @param directory Lucene index where the FASTA entries should be stored.
  * @param transformer Converts FASTA entries to documents
  */
-case class LuceneSink(directory: Directory, splitter: Splitter ,transformer: FastaEntry => Document) extends GraphStageWithMaterializedValue[SinkShape[FastaEntry], Future[List[String]]] {
+case class LuceneSink(directory: Directory, splitter: Splitter, transformer: FastaEntry => Document) extends GraphStageWithMaterializedValue[SinkShape[FastaEntry], Future[List[String]]] {
 
   val in: Inlet[FastaEntry] = Inlet("input")
