+++ subprojects/stream/src/test/resources/at/bioinform/stream/fasta/fasta_easy.fa
