+++ subprojects/stream/src/test/resources/at/bioinform/stream/fasta/fasta_with_descriptions.fa
