+++ subprojects/stream/src/test/resources/at/bioinform/stream/lucene/fasta_easy.fa
