+++ subprojects/stream/src/test/scala/at/bioinform/stream/fasta/FastaFlowTest.scala
@@ -1,4 +1,4 @@
-package at.bioinform.codec.fasta
+package at.bioinform.stream.fasta
 
 import java.nio.file.Paths
 
@@ -18,14 +18,14 @@ class FastaFlowTest extends TestKit(ActorSystem("FastaProcessorTest")) with FunS
 
   describe("A FastaFlow") {
     it("should parse an empty file.") {
-      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/codec/fasta/fasta_empty.fa").toURI))
+      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/stream/fasta/fasta_empty.fa").toURI))
         .runWith(TestSink.probe[FastaEntry])
         .request(10)
         .expectComplete()
     }
 
     it("should parse a very standard FASTA file.") {
-      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/codec/fasta/fasta_easy.fa").toURI))
+      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/stream/fasta/fasta_easy.fa").toURI))
         .runWith(TestSink.probe[FastaEntry])
         .request(2)
         .expectNext(
@@ -45,7 +45,7 @@ class FastaFlowTest extends TestKit(ActorSystem("FastaProcessorTest")) with FunS
     }
 
     it("should parse a FASTA file with empty lines.") {
-      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/codec/fasta/fasta_with_empty_lines.fa").toURI))
+      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/stream/fasta/fasta_with_empty_lines.fa").toURI))
         .runWith(TestSink.probe[FastaEntry])
         .request(2)
         .expectNext(
@@ -61,7 +61,7 @@ class FastaFlowTest extends TestKit(ActorSystem("FastaProcessorTest")) with FunS
     }
 
     it("should parse a FASTA file with comments.") {
-      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/codec/fasta/fasta_with_comments.fa").toURI))
+      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/stream/fasta/fasta_with_comments.fa").toURI))
         .runWith(TestSink.probe[FastaEntry])
         .request(2)
         .expectNext(
@@ -77,7 +77,7 @@ class FastaFlowTest extends TestKit(ActorSystem("FastaProcessorTest")) with FunS
     }
 
     it("should parse a FASTA file containing sequence descriptions.") {
-      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/codec/fasta/fasta_with_descriptions.fa").toURI))
+      FastaFlow.from(Paths.get(getClass.getResource("/at/bioinform/stream/fasta/fasta_with_descriptions.fa").toURI))
         .runWith(TestSink.probe[FastaEntry])
         .request(2)
         .expectNext(
