+++ subprojects/stream/src/test/scala/at/bioinform/stream/fasta/SplitterTest.scala
@@ -1,4 +1,4 @@
-package at.bioinform.codec.fasta
+package at.bioinform.stream.fasta
 
 import org.scalatest.{FunSpec, Matchers}
